added the order tracking, and order success page as well as some other minor fixes
added order tracking from the start page. added order success screen where you get your order tracking # after order has been placed.

+++ cartpage.php
@@ -7,16 +7,23 @@
 
 <body>
     <?php
-    # ALEXANDER KASHYAP DEC 2 2021
+    /*
+        Alexander Kashyap Z1926618
+        DEC 2 2021
+        Group Project
+    */
+
+    #including loging credentials and library
     include("creds.php");
     include("library.php");
+
     try {
         # connect to DB
         $dsn = "mysql:host=courses;dbname=z1926618";
         $pdo = new PDO($dsn, $username, $password);
 
         # Showing the cart
-        echo "<h2>Shopping Cart</h2>";
+        echo "<h1>Shopping Cart</h1>";
         $rs = $pdo->query("SELECT * FROM CART;");
         $rows = $rs->fetchAll(PDO::FETCH_ASSOC);
         draw_table($rows);
@@ -25,20 +32,77 @@
         $rs = $pdo->query("SELECT COST FROM CART;");
         $rows = $rs->fetchAll(PDO::FETCH_ASSOC);
         $total = 0;
-        foreach ($rows as $item) {
+        foreach ($rows as $item) { # sum the cost of all things in the cart
             $total = $total + $item['COST'];
         }
-        echo "<br>Current Total = " . $total;
+        echo "<br><h3>Current Total = $" . $total . "</h3>";
 
-        #clear cart vutton
+        ### Update quanity of select item from a drop down
+        echo '<br><br>Update Quantity:';
         echo '<form method="post">';
-        echo '<br><input type="submit" value="Clear Cart" />';
+        echo '<label for="parts">Item from cart:</label>';
+        echo '<select id="parts" name="parts">';
+        $rs = $pdo->query("SELECT PRODUCT_NAME FROM CART;");
+        $row = $rs->fetchAll(PDO::FETCH_ASSOC);
+        echo '<option value="">Select Part</option>';
+        foreach ($row as $item) {
+            #the options to select from in the drop down
+            echo '<option value="' . $item["PRODUCT_NAME"] . '">' . $item["PRODUCT_NAME"] . "</option>";
+        }
+        echo '</select>';
+
+        #input the new quantity for that part you want to change
+        echo '<label for="quantity">New Quantity:</label>';
+        echo '<input type="text" id="quantity" name="quantity">';
+
+        # hit the update button to update...
+        echo '<input type="submit" value="Update" />';
+        echo "</form>";
+
+        ### clear aselect item
+        echo '<br>Remove Item:';
+        echo '<form method="post">';
+        echo '<label for="rpart">Item from cart:</label>';
+        echo '<select id="rpart" name="rpart">';
+        $rs = $pdo->query("SELECT PRODUCT_NAME FROM CART;");
+        $row = $rs->fetchAll(PDO::FETCH_ASSOC);
+        echo '<option value="">Select Part</option>';
+        foreach ($row as $item) {
+            #the options to select from in the drop down
+            echo '<option value="' . $item["PRODUCT_NAME"] . '">' . $item["PRODUCT_NAME"] . "</option>";
+        }
+        echo '</select>';
+        echo '<input type="submit" value="Remove" />';
+        echo "</form>";
+
+        #clear entire cart vutton
+        echo '<form method="post">';
+        echo '<br><input type="submit" name="clear" value="Clear Entire Cart" />';
         echo "</form>";
 
         #if we've recieved a post (clear cart button has been pressed) do this...
         if ($_SERVER['REQUEST_METHOD'] === 'POST') {
             #clear the cart ...
-            $pdo->exec('delete from CART');
+            if (isset($_POST['clear'])) {
+                $pdo->exec('delete from CART');
+                #refresh page
+                echo "<meta http-equiv='refresh' content='0'>";
+            }
+            if (isset($_POST['parts'])) {
+                $pdo->exec('UPDATE CART SET QUANTITY = ' . $_POST['quantity'] . ' WHERE PRODUCT_NAME="' . $_POST['parts'] . '";');
+                $rs = $pdo->query('select PRICE from PRODUCTS where PRODUCT_NAME="' . $_POST['parts'] . '";');
+                $row = $rs->fetch(PDO::FETCH_ASSOC);
+                $newcost = $row['PRICE'] * $_POST['quantity'];
+                $pdo->exec('UPDATE CART SET COST = ' . $newcost . ' WHERE PRODUCT_NAME="' . $_POST['parts'] . '";');
+
+                #refresh page
+                echo "<meta http-equiv='refresh' content='0'>";
+            }
+            if (isset($_POST['rpart'])) {
+                $pdo->exec('delete from CART where PRODUCT_NAME="' . $_POST['rpart'] . '";');
+                #refresh
+                echo "<meta http-equiv='refresh' content='0'>";
+            }
         }
 
         #go to shopping cart screen
